Move warning notifications to own component
